Moving files to a sub-directory to make space for other things
In PR's coming next we are going to have things that aren't
matchers, so it's good to separate them out. So this doesn't
neccessarily make much sense... yet.
