+++ src/h_matchers/__init__.py
@@ -1,21 +1,3 @@
-"""The public interface class for comparing with anything."""
+"""Exposes the user facing matchers for importing"""
 
-from h_matchers.code import AnyCallable, AnyFunction, AnyInstanceOf
-from h_matchers.core import Matcher
-from h_matchers.string import AnyString
-
-__all__ = ["Any"]
-
-
-class Any(Matcher):
-    """Matches anything and provides access to other matchers."""
-
-    # pylint: disable=too-few-public-methods
-
-    string = AnyString
-    function = AnyFunction
-    callable = AnyCallable
-    instance_of = AnyInstanceOf
-
-    def __init__(self):
-        super().__init__("* anything *", lambda _: True)
+from h_matchers.matcher.any import Any
