+++ src/h_matchers/matcher/code.py
@@ -4,7 +4,7 @@
 
 from inspect import isclass
 
-from h_matchers.core import Matcher
+from h_matchers.matcher.core import Matcher
 
 __all__ = ["AnyInstanceOf", "AnyFunction"]
