+++ src/h_matchers/matcher/string.py
@@ -4,7 +4,7 @@
 
 import re
 
-from h_matchers.core import Matcher
+from h_matchers.matcher.core import Matcher
 
 __all__ = ["AnyString", "AnyStringContaining", "AnyStringMatching"]
