+++ tests/unit/h_matchers/matcher/any_test.py
@@ -1,10 +1,12 @@
 import pytest
 from tests.unit.data_types import DataTypes
 
-from h_matchers import Any
+from h_matchers.matcher.any import Any
 
 
 class TestAnything:
+    # pylint: disable=protected-access
+
     @pytest.mark.parametrize("item,_", DataTypes.parameters())
     def test_it_matches(self, item, _):
         assert Any() == item
