+++ tests/unit/h_matchers/matcher/string_test.py
@@ -3,7 +3,7 @@ import re
 import pytest
 from tests.unit.data_types import DataTypes
 
-from h_matchers.string import AnyString, AnyStringContaining, AnyStringMatching
+from h_matchers.matcher.string import AnyString, AnyStringContaining, AnyStringMatching
 
 
 class TestAnyString:
